Add composer as new service on docker-compose.yml
This repo offers to developers a docker way to work with magento and pagarme-magento module. But there was a manual process: install composer from a [pagarme-composer](https://hub.docker.com/r/pagarme/composer/) image:

```
docker run -it --rm -v .:/code -w /code install
```

This is necessary to magento works well with our module. So, this PR includes an automation for this installation process adding the configurations to `docker-compose.yml`